requirements.txt
itchat==1.3.10
requests==2.21.0
PyYAML==5.1
pydub==0.23.1
